requirements.txt
# Core FastAPI - COMPATIBLE VERSIONS
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication - COMPATIBLE VERSIONS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# TensorFlow - CLOUD COMPATIBLE
tensorflow-cpu==2.13.0
numpy==1.24.3
pillow==10.1.0

# Medical imaging
pydicom==2.4.3

# Utilities
python-dotenv==1.0.0
slowapi==0.6.0
gunicorn==21.2.0
prometheus-fastapi-instrumentator==6.1.0
